Tweaks and Fixes

+++ TroysCode/T/TScrollEvent.java
@@ -8,9 +8,9 @@ public class TScrollEvent extends AWTEvent
 
 	public static final int TSCROLLBARSCROLLED = 1001;
 	
-	public float scrollPercent;
+	public double scrollPercent;
 	
-	public TScrollEvent(Object source, int id, float scrollPercent)
+	public TScrollEvent(Object source, int id, double scrollPercent)
 		{
 			super(source, id);
 			this.scrollPercent = scrollPercent;
